requirements.txt
-i https://pypi.org/simple
amqp==5.2.0; python_version >= '3.6'
anyio==4.4.0; python_version >= '3.8'
billiard==4.2.0; python_version >= '3.7'
celery[redis]==5.4.0; python_version >= '3.8'
certifi==2024.7.4; python_version >= '3.6'
click==8.1.7; python_version >= '3.7'
click-didyoumean==0.3.1; python_full_version >= '3.6.2'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.5; python_version >= '3.8'
httpx==0.27.0; python_version >= '3.8'
idna==3.7; python_version >= '3.5'
kombu==5.3.7; python_version >= '3.8'
ollama==0.2.1; python_version >= '3.8' and python_version < '4.0'
prompt-toolkit==3.0.47; python_full_version >= '3.7.0'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.0.1; python_version >= '3.8'
redis==5.0.7
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.1; python_version >= '3.7'
tzdata==2024.1; python_version >= '2'
vine==5.1.0; python_version >= '3.6'
wcwidth==0.2.13
